+++ mobile/android/android-components/components/feature/session/src/test/java/mozilla/components/feature/session/behavior/EngineViewBottomBehaviorTest.kt
@@ -63,6 +63,8 @@ class FakeEngineView(context: Context) : TextView(context), EngineView {
 
     override fun setVerticalClipping(clippingHeight: Int) {}
 
+    override fun setDynamicToolbarMaxHeight(height: Int) {}
+
     override fun release() {}
 }